[RpcAgent] Metrics for current num active/async rpc calls.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/34398

As part of PR 34109, it was suggested that we track the number of outstanding
async calls for RPC DebugInfo, particularly if we move towards using
at::launch() threads on occasion for continuations.

This particular aspect of the change was distinct from the main purpose of the
diff, and started getting bigger, so split this functionality out as a separate diff.
For completeness, we track client_active_calls, server_active_calls,
server_active_async_calls, and write some very basic unittest coverage.
ghstack-source-id: 99701826

Differential Revision: [D20314994](https://our.internmc.facebook.com/intern/diff/D20314994/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D20314994/)!